2024-04-19 Spring Boot 3 FullCalendar

+++ fullcalendar/build/resources/main/static/js/note.js
@@ -2,17 +2,28 @@ const deleteButton = document.getElementById('delete-btn');
 const modifyButton = document.getElementById('modify-btn');
 const createButton = document.getElementById('create-btn');
 
+ClassicEditor
+    .create( document.querySelector( '#editor' ) )
+    .then( newEditor => {
+        editor = newEditor;
+    } )
+    .catch( error => {
+        console.error( error );
+    } );
 
 if(deleteButton) {
     deleteButton.addEventListener('click', event =>{
-        let id = document.getElementById('article-id').value;
-        fetch(`/api/articles/${id}`, {
-            method: 'DELETE'
-        })
-            .then(()=> {
-                alert('삭제가 완료되었습니다.');
-                location.replace('/articles');
-            });
+        let id = document.getElementById('note-id').value;
+        var confirmation = confirm("삭제하시겠습니까?");
+        if(confirmation) {
+            fetch(`/api/notes/${id}`, {
+                method: 'DELETE'
+            })
+                .then(()=> {
+                    alert('삭제가 완료되었습니다.');
+                    location.replace('/calendar');
+                });
+        }
     });
 }
 
@@ -23,7 +34,7 @@ if(modifyButton) {
         //여러개 중에 id만 갯또다제
         let id = params.get('id');
 
-        fetch(`/api/articles/${id}`, {
+        fetch(`/api/notes/${id}`, {
             method: 'PUT', //수정은 put
             headers: {
                 "Content-Type": "application/json",
@@ -35,25 +46,29 @@ if(modifyButton) {
         })
             .then(() => {
                 alert('수정이 완료되었습니다.');
-                location.replace(`/articles/${id}`);
+                location.replace(`/calendar/note/${id}`);
             });
     });
 }
 
 if(createButton) {
     createButton.addEventListener('click', (event) => {
-        fetch("/api/articles", {
+        let params = new URLSearchParams(location.search);
+        const editorData = editor.getData();
+        fetch("/api/notes", {
             method: "POST",
             headers: {
                 "Content-Type": "application/json",
             },
             body: JSON.stringify({
                 title: document.getElementById("title").value,
-                content: document.getElementById("content").value,
+                //content: document.getElementById("editor").value,
+                content: editorData,
+                start: params.get('date')
             }),
         }). then(() => {
-            alert("등록 완료되었습니다.");
-            location.replace("/articles");
+            //alert("등록 완료되었습니다.");
+            location.replace("/calendar");
         });
     });
 }
